fix styling

+++ envpool/box2d/bipedal_walker_env.h
@@ -38,8 +38,8 @@ class BipedalWalkerLidarCallback : public b2RayCastCallback {
 };
 
 class BipedalWalkerBox2dEnv {
-  const double kFPS = 50;
-  const double kScale = 30.0;
+  const float kFPS = 50;
+  const float kScale = 30.0;
   const float kMotorsTorque = 80;
   const float kSpeedHip = 4;
   const float kSpeedKnee = 6;
